Updating plugins structure
