--- META-INF/plugin.xml
@@ -1,93 +1,0 @@
-<idea-plugin version="2">
-    <id>com.ruslanpolutsygan.adderremover</id>
-    <name>Adders/Removers Generator</name>
-    <version>0.3.6</version>
-    <vendor email="ruslan.polutsygan@gmail.com">Ruslan Polutsygan</vendor>
-
-    <description>Generates adders/removers for class properties</description>
-
-    <change-notes><![CDATA[
-<h2>0.3.6</h2>
-    <li>[bugfix] Support code template without phpdoc - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/12">Issue #12</a></li>
-    <li>[enhancement] New default template for remover -
-        <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/9">Issue #9</a>
-        <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/pull/10">PR #10</a>
-    </li>
-<h2>0.3.5</h2>
-<ul>
-    <li>[bugfix] Field typehinted with just `ArrayCollection` is not recognized - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/8">Issue #8</a></li>
-</ul>
-<h2>0.3.4</h2>
-<ul>
-    <li>[enhancement] Inject code to cursor position - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/6">Issue #6</a></li>
-    <li>[enhancement] Recognize concrete Doctrine Collection implementations - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/7">Issue #7</a></li>
-</ul>
-<h2>0.3.3</h2>
-<ul>
-    <li>[bugfix] NoSuchMethodError in PhpStorm 2016.3 - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/5">Issue #5</a></li>
-</ul>
-<h2>0.3.2</h2>
-<ul>
-    <li>[enhancement] ${TYPE_HINT_FQCN} template variable introduced. ${TYPE_HINT} is not fully qualified any more - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/3">Issue #3</a></li>
-    <li>[enhancement] ${THIS_CLASS_NAME} template variable introduced - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/2">Issue #2</a></li>
-    <li>[bugfix] ClassCastException in PhpStorm 2016 - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/4">Issue #4</a></li>
-</ul>
-<h2>0.3.1</h2>
-<ul>
-    <li>[bugfix] EmptySet to TreeSet cast exception fixed - <a href="https://github.com/ruslan-polutsygan/phpstorm-adder-remover-generator-plugin/issues/1">Bug #1</a></li>
-</ul>
-<h2>0.3.0</h2>
-<ul>
-    <li>[feature] adder and remover methods generated based on file templates(settings->editor->file and code templates->other->adder/remover)</li>
-    <li>[improvement] adder/remover param name generated based on field name (users => user)</li>
-</ul>
-<h2>0.2.1</h2>
-<ul>
-    <li>[improvement] default remover implementation for array property based on array_search</li>
-    <li>[bugfix] display "generate .." actions menu items only in php class context</li>
-</ul>
-<h2>0.2.0</h2>
-<ul>
-    <li>Update: support for PhpStorm 10</li>
-</ul>
-<h2>0.1.0</h2>
-<ul>
-    <li>Init version. Basic adders / removers generators</li>
-</ul>
-      ]]>
-    </change-notes>
-
-    <idea-version since-build="131"/>
-
-    <!-- please see https://confluence.jetbrains.com/display/IDEADEV/Plugin+Compatibility+with+IntelliJ+Platform+Products
-         on how to target different products -->
-    <!-- uncomment to enable plugin in all products
-    <depends>com.intellij.modules.lang</depends>
-    -->
-    <depends>com.jetbrains.php</depends>
-    <depends>com.intellij.modules.platform</depends>
-
-    <extensions defaultExtensionNs="com.intellij">
-        <fileTemplateGroup implementation="com.ruslanpolutsygan.adderremover.ft.AdderRemoverTemplateDescriptorFactory"/>
-    </extensions>
-
-    <application-components>
-        <!-- Add your application components here -->
-    </application-components>
-
-    <project-components>
-        <!-- Add your project components here -->
-    </project-components>
-
-    <actions>
-        <action id="Php.Generators.Adders" class="com.ruslanpolutsygan.adderremover.actions.PhpGenerateAddersAction" text="Generate Adders">
-            <add-to-group group-id="GenerateGroup" anchor="last" />
-        </action>
-        <action id="Php.Generators.Removers" class="com.ruslanpolutsygan.adderremover.actions.PhpGenerateRemoversAction" text="Generate Removers">
-            <add-to-group group-id="GenerateGroup" anchor="last" />
-        </action>
-        <action id="Php.Generators.AddersAndRemovers" class="com.ruslanpolutsygan.adderremover.actions.PhpGenerateAddersAndRemoversAction" text="Generate Adders and Removers">
-            <add-to-group group-id="GenerateGroup" anchor="last" />
-        </action>
-    </actions>
-</idea-plugin>
