+++ build.gradle.kts
@@ -1,0 +1,43 @@
+plugins {
+    id("java")
+    id("org.jetbrains.intellij") version "1.10.1"
+}
+
+group = "com.johndaligault"
+version = "1.0-SNAPSHOT"
+
+repositories {
+    mavenCentral()
+}
+
+// Configure Gradle IntelliJ Plugin
+// Read more: https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html
+intellij {
+    version.set("2022.2")
+    type.set("PS") // Target IDE Platform
+
+    plugins.set(listOf("com.jetbrains.php"))
+}
+
+tasks {
+    // Set the JVM compatibility versions
+    withType<JavaCompile> {
+        sourceCompatibility = "11"
+        targetCompatibility = "11"
+    }
+
+    patchPluginXml {
+        sinceBuild.set("221")
+        untilBuild.set("231.*")
+    }
+
+    signPlugin {
+        certificateChain.set(System.getenv("CERTIFICATE_CHAIN"))
+        privateKey.set(System.getenv("PRIVATE_KEY"))
+        password.set(System.getenv("PRIVATE_KEY_PASSWORD"))
+    }
+
+    publishPlugin {
+        token.set(System.getenv("PUBLISH_TOKEN"))
+    }
+}
