--- resources/fileTemplates/j2ee/Adder.php.ft
@@ -1,13 +1,0 @@
-/**
- * @param ${TYPE_HINT} ${PARAM_NAME}
- */
-public function ${METHOD_NAME}(${TYPE_HINTED_PARAM})
-{
-#if (${IS_DOCTRINE_COLLECTION} == "doctrine")
-    $this->${FIELD_NAME}->add(${PARAM_NAME});
-    // uncomment if you want to update other side
-    //${PARAM_NAME}->set${THIS_CLASS_NAME}($this);
-#else
-    $this->${FIELD_NAME}[] = ${PARAM_NAME};
-#end
-}
