--- resources/fileTemplates/j2ee/Remover.php.ft
@@ -1,15 +1,0 @@
-/**
- * @param ${TYPE_HINT} ${PARAM_NAME}
- */
-public function ${METHOD_NAME}(${TYPE_HINTED_PARAM})
-{
-#if (${IS_DOCTRINE_COLLECTION} == "doctrine")
-    $this->${FIELD_NAME}->removeElement(${PARAM_NAME});
-    // uncomment if you want to update other side
-    //${PARAM_NAME}->set${THIS_CLASS_NAME}(null);
-#else
-    if (false !== $key = array_search(${PARAM_NAME}, $this->${FIELD_NAME}, true)) {
-        array_splice($this->${FIELD_NAME}, $key, 1);
-    }
-#end
-}
