--- src/com/ruslanpolutsygan/adderremover/handler/checkers/ChainedChecker.java
@@ -1,23 +1,0 @@
-package com.ruslanpolutsygan.adderremover.handler.checkers;
-
-import com.jetbrains.php.lang.psi.elements.Field;
-
-public class ChainedChecker extends MethodChecker {
-
-    private MethodChecker[] checkers;
-
-    public ChainedChecker(MethodChecker[] checkers) {
-        this.checkers = checkers;
-    }
-
-    @Override
-    public boolean hasMethod(Field field) {
-        for(MethodChecker checker : this.checkers) {
-            if(checker.hasMethod(field)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-}
