+++ src/main/java/com/johndaligault/adderremover/handler/generators/MethodGenerator.java
@@ -1,0 +1,8 @@
+package com.johndaligault.adderremover.handler.generators;
+
+import com.jetbrains.php.lang.psi.elements.Field;
+import com.johndaligault.adderremover.handler.PhpMethodData;
+
+public abstract class MethodGenerator {
+    public abstract PhpMethodData[] generate(Field field);
+}
